requirements.txt
python-telegram-bot==20.3
requests==2.28.2
beautifulsoup4==4.11.1
python-dotenv==1.0.0
aiohttp==3.8.5
google-api-python-client==2.70.0
google-auth==2.16.0
openai==1.3.0
